Completed Title Screen.

+++ Community Game Jam 2k19/Assets/Scripts/InterruptionObject.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 public class InterruptionObject : ScriptableObject
 {
     public float speed;
-    public int difficulty;
+    public float difficulty;
 }
